Separate lib code into core module
Made a start on splitting the code up. Currently moved State and App
into a core module with their own modules each might be overkill? Still
need to extract out the shader and pipeline handling so it can hopefully
be more dynamic.
